Adding SAT method of validation
This adds a major new method of validation for LC's.  We include the dependency sat.js by Gregory Duck, and then add the necessary infrastructure to convert the validation of and LC turnstyle into the input format expected by satSolve().  The majority of the code (and execution time) is in converting the LC's to the appropriate arrays of integers.  We use the method of introducing switch variables to reduce this process from exponential to quadratic growth.   This code is not yet optimized, but does work.  Optimization will come later.

We also include the dependencies and some related references.