SAT Validation 2.0
Complete reimplementation of SAT Validation to improve efficiency.  Converts to .cnf() which has atoms that are toString() of statements and switch variables.  Converts to satcnf() to convert such cnfs to the form needed for satSolve(). Renames the previous Validate() routine using the 1.0 SAT implementation to oldValidate() and replaces it with Validate() using the new more efficient method.  Note that oldValidate() still works, and uses .CNF() instead of .cnf().

There is also a mathematical difference between the two validtion routines.  In particular, trailing Givens inside an environment are now ignored, whereas in the first implementation they were not.

We also add a new testing suite to put the new routines through their paces.